backend/requirements.txt
# Core Framework
fastapi[all]==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
alembic==1.13.1
psycopg[binary,pool]==3.2.1
asyncpg==0.29.0

# Authentication
fastapi-users[sqlalchemy]==13.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caching & Jobs
redis==5.0.1
celery[redis]==5.4.0

# HTTP & Scraping
httpx==0.27.0
aiohttp==3.10.5
beautifulsoup4==4.12.3
lxml==5.3.0
readability-lxml==0.8.1
html2text==2024.2.26
# playwright==1.47.0  # Will add when porting Article extractor

# Social Media
yt-dlp==2024.8.6

# Email
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0

# LLM Providers
openai==1.51.0
anthropic==0.39.0
google-generativeai==0.8.2
google-genai==0.2.2

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
tenacity==9.0.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.8.0
ruff==0.6.8
mypy==1.11.2